Fix svg_attr vprint issue properly
Abandoned the use of vprintf in svg_atter since using a va_list after a
vprintf call is undefined, and GCC and MinGW/Win use differenet
defualts. Replaced svg_attr macro with a function that returns a sting
allocated for the attribute. The downside is the string must be free'd
later in the program. In this case, all svg_attr return strings are
free'd in the _svg_tag function.